@@ -35,14 +35,14 @@
     "url": "https://github.com/danielmoncada/date-time-picker.git"
   },
   "dependencies": {
-    "@angular/animations": "^13.3.5",
-    "@angular/common": "^13.1.2",
-    "@angular/compiler": "^13.3.5",
-    "@angular/core": "^13.1.2",
-    "@angular/forms": "^13.3.5",
-    "@angular/platform-browser": "^13.3.5",
-    "@angular/platform-browser-dynamic": "^13.3.5",
-    "@angular/router": "^13.3.5",
+    "@angular/animations": "^14.3.0",
+    "@angular/common": "^14.3.0",
+    "@angular/compiler": "^14.3.0",
+    "@angular/core": "^14.3.0",
+    "@angular/forms": "^14.3.0",
+    "@angular/platform-browser": "^14.3.0",
+    "@angular/platform-browser-dynamic": "^14.3.0",
+    "@angular/router": "^14.3.0",
     "core-js": "^3.22.3",
     "moment": "^2.29.3",
     "moment-timezone": "^0.5.34",
@@ -53,11 +53,11 @@
     "zone.js": "~0.11.5"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^13.3.4",
+    "@angular-devkit/build-angular": "^14.2.13",
     "@angular/cdk": "^13.3.5",
-    "@angular/cli": "^13.3.4",
-    "@angular/compiler-cli": "^13.3.5",
-    "@angular/language-service": "^13.3.5",
+    "@angular/cli": "^14.2.13",
+    "@angular/compiler-cli": "^14.3.0",
+    "@angular/language-service": "^14.3.0",
     "@types/jasmine": "^3.10.6",
     "@types/jasminewd2": "^2.0.10",
     "@types/node": "^12.20.50",
@@ -70,16 +70,14 @@
     "karma-coverage": "^2.2.0",
     "karma-jasmine": "~5.0.0",
     "karma-jasmine-html-reporter": "^1.7.0",
-    "ng-packagr": "^13.3.1",
+    "ng-packagr": "^14.2.2",
     "node-sass": "^9.0.0",
     "protractor": "~7.0.0",
     "ts-node": "^8.10.2",
     "tslint": "^6.1.3",
-    "typescript": "~4.5.4"
+    "typescript": "~4.8.4"
   },
   "peerDependencies": {
-    "@angular/common": "^13.0.3 || ^14.0.0 || ^15.0.0 || ^16.0.0",
-    "@angular/core": "^13.0.3 || ^14.0.0 || ^15.0.0 || ^16.0.0",
     "@angular/cdk": "^13.0.3 || ^14.0.0 || ^15.0.0 || ^16.0.0"
   }
 }
